updates/2024-12-3/requirements.txt
pymongo==4.6.1
python-dotenv==1.0.0